+++ spec/wechat_pay/native_spec.rb
@@ -1,0 +1,10 @@
+require 'spec_helper'
+
+describe WechatPay::Native do
+  it ".payment" do
+    product_id = 'test_product_id'
+    url = WechatPay::Native.payment(product_id)
+    url.must_include 'weixin://wxpay/bizpayurl?'
+    url.must_include product_id
+  end
+end